Use ICU4J specific override of the initialize-package-file function.
The default initialize-package-file function will strip all directory
entries from JAR files. That does not work well with ICU, which in many
places depends on directory entries being present in order to find data
in JARs. Therefore the default function is overridden here to not strip
any directory entries.

(It would be nice if the build core definitions instead were enhanced to
make such behaviour easily configured, without the need for such a hack
as this function override.)

bug: 19108324
Change-Id: I33e9dfb5510c7b5a2411d85480feeb051f5ebb42