The initial check-in of some ICU4J src transformation tools
Adds the bones of a generic library for source code manipulation
(currysrc) and associated build rules.

Adds ICU4J-specific rules classes for generating an
android_icu4j directory and a script to invoke it.

Includes the generated android_icu4j directory.

No real documentation transformations are included right now besides
the addition of a source-file header indicating the origin of the
generated source code. The aim of this check-in is to replace jarjar
in the core-libart build process.

Bug: 22023363
Change-Id: I138149956dc66448839b44f065615b9268ec7f76