Improvements to the currysrc / srcgen tools
Improvements to currysrc to enable AstTransformer and DocumentTransformer to
be interleaved. Previously Main assumed AST changes were followed by
Document changes. Some Javadoc changes like removing tags seem easier when
done as an AstTransformer so it helped if they can be applied in any order.

Modified the ICU src to:
1) Hack around some badly placed javadoc tags in Bidi
2) Remove @stable tags generally
3) Account for changes in currysrc

Bug: 22023363
Change-Id: If9b4721b4ce85c0d68030fd5fc03f2b7d5cd76f4