Hide static setDefault methods in ICU4J
These methods interact with java.util.TimeZone and
java.util.Locale. Their interactions have not been properly
tested so they are being hidden for now.

Some related constants have also been hidden.

Some comments need to be updated as a result of this:
replacement comments have been added to replacements.txt.

This change also fixes a bug in MethodLocator: the type
was not being checked so it was matching any method with the
same name / parameters regardless of type.

Bug: 25924145
Change-Id: I3fea5e3d88417cfb702550147c0c33d1cd2d0dc9