Patch for SoftCache to explicitly support null values
SoftCache now explicitly stores null values (rather than treating
them like values that have been collected). This is useful if null is
a valid cacheable value: the null will not be treated as a miss,
which would involve "recalculating" the null.

This change also lays the groundwork for references that may not be
SoftReferences (the actual change there left for a follow up).

Accessibility of nested class methods modified to avoid synthetic
methods.

Bug: 28326526
Change-Id: I9111984302b18fdc594ea3d2a8c39004b1706ba6